refactor(guest-agent): split session history reader tests
