+++ crates/guest-agent/tests/codex_session_resume.rs
@@ -1,4 +1,4 @@
-//! Integration tests for codex session-resume + claude session-history.
+//! Integration tests for codex session-resume metadata capture.
 //!
 //! # Why a dedicated test binary
 //!
@@ -18,13 +18,8 @@
 //!
 //! - End-to-end `send_event` → `capture_session_metadata` → marker write for the
 //!   codex `thread.started` event shape.
-//! - `session_history::read_session_history` decodes the marker and walks the
-//!   `YYYY/MM/DD/` codex layout.
-//! - Claude literal-path resolution through the same public entry,
-//!   exercised by passing a literal `.jsonl` path through the
-//!   history-path file.
-//! - Negative path: a codex marker pointing at an empty sessions dir
-//!   surfaces the "file not found" error rather than a silent fallback.
+//! - Marker repair after a partial metadata write.
+//! - Invalid/non-Codex events do not persist Codex session metadata.
 
 mod common;
 
@@ -89,27 +84,6 @@ fn send_event_for_test(
 fn reset_session_files() {
     let _ = std::fs::remove_file(guest_agent::paths::session_id_file());
     let _ = std::fs::remove_file(guest_agent::paths::session_history_path_file());
-}
-
-/// Build a `YYYY/MM/DD/` style nested path under `root` and write a file.
-///
-/// Returns `Result<_, String>` rather than `unwrap`-ing because clippy's
-/// `allow-unwrap-in-tests` only whitelists `#[test]` bodies, not module-
-/// level helpers. Test bodies forward via `?`.
-fn write_session_file(
-    root: &Path,
-    sub: &[&str],
-    filename: &str,
-    content: &[u8],
-) -> Result<(), String> {
-    let mut dir = root.to_path_buf();
-    for s in sub {
-        dir.push(s);
-    }
-    std::fs::create_dir_all(&dir).map_err(|e| format!("create_dir_all {}: {e}", dir.display()))?;
-    let path = dir.join(filename);
-    std::fs::write(&path, content).map_err(|e| format!("write {}: {e}", path.display()))?;
-    Ok(())
 }
 
 #[test]
@@ -301,618 +275,4 @@ fn send_event_codex_ignores_malformed_thread_id() {
             assert_eq!(masker.mask_string(thread_id), thread_id);
         }
     }
-}
-
-#[test]
-fn read_session_history_resolves_codex_marker_end_to_end() {
-    setup_env_once();
-    let _guard = TEST_MUTEX.lock().unwrap();
-    reset_session_files();
-
-    // Set up a fake codex sessions tree under a temp dir, with a jsonl file
-    // at the canonical YYYY/MM/DD/ depth.
-    let tmp = tempfile::tempdir().unwrap();
-    let sessions_dir = tmp.path().join("sessions");
-    let thread_id = "0193abcd-ef01-7234-89ab-cdef01234567";
-    let history = b"{\"type\":\"thread.started\",\"thread_id\":\"x\"}\n";
-    write_session_file(
-        &sessions_dir,
-        &["2026", "04", "28"],
-        &format!("{thread_id}.jsonl"),
-        history,
-    )
-    .unwrap();
-
-    // Drive the public entry exactly the way `checkpoint.rs` does:
-    // a marker file pointing at the sessions dir + thread id.
-    let path_file = tmp.path().join("path.txt");
-    let marker = format!(
-        "CODEX_SEARCH:{}:{thread_id}",
-        sessions_dir.to_string_lossy()
-    );
-    std::fs::write(&path_file, marker.as_bytes()).unwrap();
-
-    let bytes =
-        guest_agent::session_history::read_session_history(path_file.to_str().unwrap()).unwrap();
-    assert_eq!(bytes, history);
-}
-
-#[test]
-fn read_session_history_decodes_legacy_zstd_session() {
-    setup_env_once();
-    let _guard = TEST_MUTEX.lock().unwrap();
-    reset_session_files();
-
-    let tmp = tempfile::tempdir().unwrap();
-    let sessions_dir = tmp.path().join("sessions");
-    let thread_id = "0193abcd-ef01-7234-89ab-cdef01234567";
-    let history = b"{\"type\":\"thread.started\",\"thread_id\":\"x\"}\n";
-    let compressed = zstd::encode_all(history.as_slice(), 0).unwrap();
-    write_session_file(
-        &sessions_dir,
-        &["2026", "04", "28"],
-        &format!("{thread_id}.jsonl.zst"),
-        &compressed,
-    )
-    .unwrap();
-
-    let path_file = tmp.path().join("path.txt");
-    let marker = format!(
-        "CODEX_SEARCH:{}:{thread_id}",
-        sessions_dir.to_string_lossy()
-    );
-    std::fs::write(&path_file, marker.as_bytes()).unwrap();
-
-    let bytes =
-        guest_agent::session_history::read_session_history(path_file.to_str().unwrap()).unwrap();
-    assert_eq!(bytes, history);
-}
-
-#[test]
-fn read_session_history_rejects_duplicate_codex_matches() {
-    setup_env_once();
-    let _guard = TEST_MUTEX.lock().unwrap();
-    reset_session_files();
-
-    let tmp = tempfile::tempdir().unwrap();
-    let sessions_dir = tmp.path().join("sessions");
-    let thread_id = "0193abcd-ef01-7234-89ab-cdef01234567";
-    write_session_file(
-        &sessions_dir,
-        &["2026", "04", "28"],
-        &format!("{thread_id}.jsonl"),
-        b"first\n",
-    )
-    .unwrap();
-    write_session_file(
-        &sessions_dir,
-        &["2026", "04", "29"],
-        &format!("rollout-2026-04-29T11-22-37-{thread_id}.jsonl"),
-        b"second\n",
-    )
-    .unwrap();
-
-    let path_file = tmp.path().join("path.txt");
-    let marker = format!(
-        "CODEX_SEARCH:{}:{thread_id}",
-        sessions_dir.to_string_lossy()
-    );
-    std::fs::write(&path_file, marker.as_bytes()).unwrap();
-
-    let err = guest_agent::session_history::read_session_history(path_file.to_str().unwrap())
-        .expect_err("duplicate codex sessions must fail clearly");
-    let msg = err.to_string();
-    assert!(
-        msg.contains("Multiple Codex session files found"),
-        "expected duplicate-session error, got: {msg}"
-    );
-    assert!(
-        !msg.contains(thread_id),
-        "duplicate-session error must not expose thread id, got: {msg}"
-    );
-}
-
-#[test]
-fn read_session_history_rejects_duplicate_jsonl_and_zstd_matches() {
-    setup_env_once();
-    let _guard = TEST_MUTEX.lock().unwrap();
-    reset_session_files();
-
-    let tmp = tempfile::tempdir().unwrap();
-    let sessions_dir = tmp.path().join("sessions");
-    let thread_id = "0193abcd-ef01-7234-89ab-cdef01234567";
-    write_session_file(
-        &sessions_dir,
-        &["2026", "04", "28"],
-        &format!("{thread_id}.jsonl"),
-        b"jsonl\n",
-    )
-    .unwrap();
-    let compressed = zstd::encode_all(b"zstd\n".as_slice(), 0).unwrap();
-    write_session_file(
-        &sessions_dir,
-        &["2026", "04", "29"],
-        &format!("{thread_id}.jsonl.zst"),
-        &compressed,
-    )
-    .unwrap();
-
-    let path_file = tmp.path().join("path.txt");
-    let marker = format!(
-        "CODEX_SEARCH:{}:{thread_id}",
-        sessions_dir.to_string_lossy()
-    );
-    std::fs::write(&path_file, marker.as_bytes()).unwrap();
-
-    let err = guest_agent::session_history::read_session_history(path_file.to_str().unwrap())
-        .expect_err("duplicate codex sessions must include zstd matches");
-    let msg = err.to_string();
-    assert!(
-        msg.contains("Multiple Codex session files found"),
-        "expected duplicate-session error, got: {msg}"
-    );
-    assert!(
-        !msg.contains(thread_id),
-        "duplicate-session error must not expose thread id, got: {msg}"
-    );
-}
-
-#[test]
-fn read_session_history_rejects_duplicate_before_reading_corrupt_zstd() {
-    setup_env_once();
-    let _guard = TEST_MUTEX.lock().unwrap();
-    reset_session_files();
-
-    let tmp = tempfile::tempdir().unwrap();
-    let sessions_dir = tmp.path().join("sessions");
-    let thread_id = "0193abcd-ef01-7234-89ab-cdef01234567";
-    write_session_file(
-        &sessions_dir,
-        &["2026", "04", "28"],
-        &format!("{thread_id}.jsonl.zst"),
-        b"not zstd",
-    )
-    .unwrap();
-    write_session_file(
-        &sessions_dir,
-        &["2026", "04", "29"],
-        &format!("rollout-2026-04-29T11-22-37-{thread_id}.jsonl.zst"),
-        b"also not zstd",
-    )
-    .unwrap();
-
-    let path_file = tmp.path().join("path.txt");
-    let marker = format!(
-        "CODEX_SEARCH:{}:{thread_id}",
-        sessions_dir.to_string_lossy()
-    );
-    std::fs::write(&path_file, marker.as_bytes()).unwrap();
-
-    let err = guest_agent::session_history::read_session_history(path_file.to_str().unwrap())
-        .expect_err("duplicate codex sessions must fail before decoding any candidate");
-    let msg = err.to_string();
-    assert!(
-        msg.contains("Multiple Codex session files found"),
-        "expected duplicate-session error, got: {msg}"
-    );
-    assert!(
-        !msg.contains(thread_id),
-        "duplicate-session error must not expose thread id, got: {msg}"
-    );
-}
-
-#[test]
-fn read_session_history_corrupt_zstd_error_omits_thread_id() {
-    setup_env_once();
-    let _guard = TEST_MUTEX.lock().unwrap();
-    reset_session_files();
-
-    let tmp = tempfile::tempdir().unwrap();
-    let sessions_dir = tmp.path().join("sessions");
-    let thread_id = "0193abcd-ef01-7234-89ab-cdef01234567";
-    write_session_file(
-        &sessions_dir,
-        &["2026", "04", "28"],
-        &format!("{thread_id}.jsonl.zst"),
-        b"not zstd",
-    )
-    .unwrap();
-
-    let path_file = tmp.path().join("path.txt");
-    let marker = format!(
-        "CODEX_SEARCH:{}:{thread_id}",
-        sessions_dir.to_string_lossy()
-    );
-    std::fs::write(&path_file, marker.as_bytes()).unwrap();
-
-    let err = guest_agent::session_history::read_session_history(path_file.to_str().unwrap())
-        .expect_err("corrupt zstd codex session must fail clearly");
-    let msg = err.to_string();
-    assert!(
-        msg.contains("Failed to decompress zstd session history"),
-        "expected zstd decode error, got: {msg}"
-    );
-    assert!(
-        !msg.contains(thread_id),
-        "zstd decode error must not expose thread id, got: {msg}"
-    );
-}
-
-#[test]
-fn read_session_history_resolves_dash_stripped_filename() {
-    // Real codex CLI prefixes filenames with `rollout-{ts}-` and the
-    // concatenation strips the UUID dashes — the substring matcher must
-    // handle that. Bug-prone enough to deserve its own integration case.
-    setup_env_once();
-    let _guard = TEST_MUTEX.lock().unwrap();
-    reset_session_files();
-
-    let tmp = tempfile::tempdir().unwrap();
-    let sessions_dir = tmp.path().join("sessions");
-    let thread_id = "0193abcd-ef01-7234-89ab-cdef01234567";
-    let id_no_dashes = thread_id.replace('-', "");
-    let history = b"line1\nline2\n";
-    write_session_file(
-        &sessions_dir,
-        &["2026", "04", "28"],
-        &format!("rollout-2026-04-28T11-22-37-{id_no_dashes}.jsonl"),
-        history,
-    )
-    .unwrap();
-
-    let path_file = tmp.path().join("path.txt");
-    let marker = format!(
-        "CODEX_SEARCH:{}:{thread_id}",
-        sessions_dir.to_string_lossy()
-    );
-    std::fs::write(&path_file, marker.as_bytes()).unwrap();
-
-    let bytes =
-        guest_agent::session_history::read_session_history(path_file.to_str().unwrap()).unwrap();
-    assert_eq!(bytes, history);
-}
-
-#[test]
-fn read_session_history_codex_marker_with_no_match_fails_fast() {
-    // Verifies the post-fix behaviour (#11430 review feedback): when no
-    // filename matches the dash-stripped UUID, return a "not found"
-    // error instead of silently picking some unrelated recent file.
-    setup_env_once();
-    let _guard = TEST_MUTEX.lock().unwrap();
-    reset_session_files();
-
-    let tmp = tempfile::tempdir().unwrap();
-    let sessions_dir = tmp.path().join("sessions");
-    // Plant an unrelated file under the tree so any silent fallback
-    // would have something to pick. The fix-fast path must reject it.
-    write_session_file(
-        &sessions_dir,
-        &["2026", "04", "27"],
-        "rollout-unrelated.jsonl.zst",
-        &zstd::encode_all(b"unrelated".as_slice(), 0).unwrap(),
-    )
-    .unwrap();
-
-    let path_file = tmp.path().join("path.txt");
-    let unknown_id = "ffffffff-ffff-7fff-bfff-ffffffffffff";
-    let marker = format!(
-        "CODEX_SEARCH:{}:{unknown_id}",
-        sessions_dir.to_string_lossy()
-    );
-    std::fs::write(&path_file, marker.as_bytes()).unwrap();
-
-    let err = guest_agent::session_history::read_session_history(path_file.to_str().unwrap())
-        .expect_err("missing codex session must surface as an error");
-    let msg = err.to_string();
-    assert!(
-        msg.contains("Codex session file not found"),
-        "expected fail-fast error, got: {msg}"
-    );
-    assert!(
-        !msg.contains(unknown_id),
-        "missing-session error must not expose thread id, got: {msg}"
-    );
-}
-
-#[test]
-fn read_session_history_codex_marker_rejects_dash_only_thread_id() {
-    setup_env_once();
-    let _guard = TEST_MUTEX.lock().unwrap();
-    reset_session_files();
-
-    let tmp = tempfile::tempdir().unwrap();
-    let sessions_dir = tmp.path().join("sessions");
-    write_session_file(
-        &sessions_dir,
-        &["2026", "04", "27"],
-        "rollout-unrelated.jsonl",
-        b"unrelated\n",
-    )
-    .unwrap();
-
-    let path_file = tmp.path().join("path.txt");
-    let marker = format!("CODEX_SEARCH:{}:---", sessions_dir.to_string_lossy());
-    std::fs::write(&path_file, marker.as_bytes()).unwrap();
-
-    let err = guest_agent::session_history::read_session_history(path_file.to_str().unwrap())
-        .expect_err("dash-only codex thread id must not match every history file");
-    let msg = err.to_string();
-    assert!(
-        msg.contains("Codex session file not found"),
-        "expected malformed thread id to fail fast, got: {msg}"
-    );
-}
-
-#[test]
-fn read_session_history_codex_marker_rejects_short_thread_id() {
-    setup_env_once();
-    let _guard = TEST_MUTEX.lock().unwrap();
-    reset_session_files();
-
-    let tmp = tempfile::tempdir().unwrap();
-    let sessions_dir = tmp.path().join("sessions");
-    write_session_file(
-        &sessions_dir,
-        &["2026", "04", "27"],
-        "rollout-abc.jsonl",
-        b"unrelated\n",
-    )
-    .unwrap();
-
-    let path_file = tmp.path().join("path.txt");
-    let marker = format!("CODEX_SEARCH:{}:abc", sessions_dir.to_string_lossy());
-    std::fs::write(&path_file, marker.as_bytes()).unwrap();
-
-    let err = guest_agent::session_history::read_session_history(path_file.to_str().unwrap())
-        .expect_err("short codex thread id must not match unrelated history files");
-    let msg = err.to_string();
-    assert!(
-        msg.contains("Codex session file not found"),
-        "expected malformed thread id to fail fast, got: {msg}"
-    );
-}
-
-#[test]
-fn read_session_history_read_error_omits_literal_session_path() {
-    setup_env_once();
-    let _guard = TEST_MUTEX.lock().unwrap();
-    reset_session_files();
-
-    let tmp = tempfile::tempdir().unwrap();
-    let session_id = "sess-secret-123";
-    let history = tmp.path().join(format!("{session_id}.jsonl"));
-    let path_file = tmp.path().join("path.txt");
-    std::fs::write(&path_file, history.to_string_lossy().as_bytes()).unwrap();
-
-    let err = guest_agent::session_history::read_session_history(path_file.to_str().unwrap())
-        .expect_err("missing literal session history path must fail clearly");
-    let msg = err.to_string();
-    assert!(
-        msg.contains("Failed to read session history"),
-        "expected read error, got: {msg}"
-    );
-    assert!(
-        !msg.contains(session_id),
-        "history read error must not expose session id, got: {msg}"
-    );
-}
-
-#[cfg(unix)]
-#[test]
-fn read_session_history_codex_marker_skips_symlinks() {
-    use std::os::unix::fs::symlink;
-
-    setup_env_once();
-    let _guard = TEST_MUTEX.lock().unwrap();
-    reset_session_files();
-
-    let tmp = tempfile::tempdir().unwrap();
-    let sessions_dir = tmp.path().join("sessions");
-    let outside_dir = tmp.path().join("outside");
-    std::fs::create_dir_all(&sessions_dir).unwrap();
-    std::fs::create_dir_all(&outside_dir).unwrap();
-
-    let thread_id = "0193abcd-ef01-7234-89ab-cdef01234567";
-    let outside_history = outside_dir.join(format!("{thread_id}.jsonl"));
-    std::fs::write(&outside_history, b"outside-history\n").unwrap();
-
-    symlink(&outside_dir, sessions_dir.join("linked-outside")).unwrap();
-    symlink(
-        &outside_history,
-        sessions_dir.join(format!("{thread_id}.jsonl")),
-    )
-    .unwrap();
-    symlink(
-        "/definitely/missing/codex-history.jsonl",
-        sessions_dir.join("dangling.jsonl"),
-    )
-    .unwrap();
-
-    let path_file = tmp.path().join("path.txt");
-    let marker = format!(
-        "CODEX_SEARCH:{}:{thread_id}",
-        sessions_dir.to_string_lossy()
-    );
-    std::fs::write(&path_file, marker.as_bytes()).unwrap();
-
-    let err = guest_agent::session_history::read_session_history(path_file.to_str().unwrap())
-        .expect_err("codex lookup must not follow symlinked history paths");
-    let msg = err.to_string();
-    assert!(
-        msg.contains("Codex session file not found"),
-        "expected symlinked candidates to be ignored, got: {msg}"
-    );
-}
-
-#[cfg(unix)]
-#[test]
-fn read_session_history_codex_marker_rejects_symlinked_sessions_root() {
-    use std::os::unix::fs::symlink;
-
-    setup_env_once();
-    let _guard = TEST_MUTEX.lock().unwrap();
-    reset_session_files();
-
-    let tmp = tempfile::tempdir().unwrap();
-    let real_sessions_dir = tmp.path().join("real-sessions");
-    let sessions_link = tmp.path().join("sessions-link");
-    std::fs::create_dir_all(&real_sessions_dir).unwrap();
-
-    let thread_id = "0193abcd-ef01-7234-89ab-cdef01234567";
-    std::fs::write(
-        real_sessions_dir.join(format!("{thread_id}.jsonl")),
-        b"outside-root-history\n",
-    )
-    .unwrap();
-    symlink(&real_sessions_dir, &sessions_link).unwrap();
-
-    let path_file = tmp.path().join("path.txt");
-    let marker = format!(
-        "CODEX_SEARCH:{}:{thread_id}",
-        sessions_link.to_string_lossy()
-    );
-    std::fs::write(&path_file, marker.as_bytes()).unwrap();
-
-    let err = guest_agent::session_history::read_session_history(path_file.to_str().unwrap())
-        .expect_err("codex lookup must not follow a symlinked sessions root");
-    let msg = err.to_string();
-    assert!(
-        msg.contains("Codex session file not found"),
-        "expected symlinked sessions root to be ignored, got: {msg}"
-    );
-}
-
-#[cfg(unix)]
-#[test]
-fn read_session_history_codex_marker_rejects_symlinked_codex_home_parent() {
-    use std::os::unix::fs::symlink;
-
-    setup_env_once();
-    let _guard = TEST_MUTEX.lock().unwrap();
-    reset_session_files();
-
-    let tmp = tempfile::tempdir().unwrap();
-    let real_codex_home = tmp.path().join("real-codex-home");
-    let codex_home_link = tmp.path().join(".codex");
-    let real_sessions_dir = real_codex_home.join("sessions");
-    std::fs::create_dir_all(&real_sessions_dir).unwrap();
-
-    let thread_id = "0193abcd-ef01-7234-89ab-cdef01234567";
-    std::fs::write(
-        real_sessions_dir.join(format!("{thread_id}.jsonl")),
-        b"outside-parent-history\n",
-    )
-    .unwrap();
-    symlink(&real_codex_home, &codex_home_link).unwrap();
-
-    let path_file = tmp.path().join("path.txt");
-    let marker = format!(
-        "CODEX_SEARCH:{}:{thread_id}",
-        codex_home_link.join("sessions").to_string_lossy()
-    );
-    std::fs::write(&path_file, marker.as_bytes()).unwrap();
-
-    let err = guest_agent::session_history::read_session_history(path_file.to_str().unwrap())
-        .expect_err("codex lookup must not follow a symlinked codex home parent");
-    let msg = err.to_string();
-    assert!(
-        msg.contains("Codex session file not found"),
-        "expected symlinked codex home parent to be ignored, got: {msg}"
-    );
-}
-
-#[cfg(unix)]
-#[test]
-fn read_session_history_codex_marker_skips_special_files() {
-    use std::os::unix::net::UnixListener;
-
-    setup_env_once();
-    let _guard = TEST_MUTEX.lock().unwrap();
-    reset_session_files();
-
-    let tmp = tempfile::tempdir().unwrap();
-    let sessions_dir = tmp.path().join("sessions");
-    std::fs::create_dir_all(&sessions_dir).unwrap();
-
-    let thread_id = "0193abcd-ef01-7234-89ab-cdef01234567";
-    let _socket = UnixListener::bind(sessions_dir.join(format!("{thread_id}.jsonl"))).unwrap();
-
-    let path_file = tmp.path().join("path.txt");
-    let marker = format!(
-        "CODEX_SEARCH:{}:{thread_id}",
-        sessions_dir.to_string_lossy()
-    );
-    std::fs::write(&path_file, marker.as_bytes()).unwrap();
-
-    let err = guest_agent::session_history::read_session_history(path_file.to_str().unwrap())
-        .expect_err("codex lookup must ignore matching non-regular files");
-    let msg = err.to_string();
-    assert!(
-        msg.contains("Codex session file not found"),
-        "expected matching special file to be ignored, got: {msg}"
-    );
-}
-
-#[cfg(unix)]
-#[test]
-fn read_session_history_codex_marker_reports_unreadable_directory() {
-    use std::os::unix::fs::PermissionsExt;
-
-    // SAFETY: `geteuid` only reads the current process credential.
-    if unsafe { libc::geteuid() } == 0 {
-        return;
-    }
-
-    setup_env_once();
-    let _guard = TEST_MUTEX.lock().unwrap();
-    reset_session_files();
-
-    let tmp = tempfile::tempdir().unwrap();
-    let sessions_dir = tmp.path().join("sessions");
-    let blocked_dir = sessions_dir.join("blocked");
-    std::fs::create_dir_all(&blocked_dir).unwrap();
-    std::fs::set_permissions(&blocked_dir, std::fs::Permissions::from_mode(0o000)).unwrap();
-
-    let thread_id = "0193abcd-ef01-7234-89ab-cdef01234567";
-    let path_file = tmp.path().join("path.txt");
-    let marker = format!(
-        "CODEX_SEARCH:{}:{thread_id}",
-        sessions_dir.to_string_lossy()
-    );
-    std::fs::write(&path_file, marker.as_bytes()).unwrap();
-
-    let result = guest_agent::session_history::read_session_history(path_file.to_str().unwrap());
-    std::fs::set_permissions(&blocked_dir, std::fs::Permissions::from_mode(0o700)).unwrap();
-
-    let err = result.expect_err("unreadable codex directories must surface as read errors");
-    let msg = err.to_string();
-    assert!(
-        msg.contains("Failed to read session history"),
-        "expected directory read error, got: {msg}"
-    );
-    assert!(
-        msg.contains("Permission denied"),
-        "expected permission failure to be preserved, got: {msg}"
-    );
-}
-
-#[test]
-fn read_session_history_resolves_claude_literal_path() {
-    // Claude path goes through the same public entry but uses a literal
-    // jsonl path rather than a marker. Covered here because the Claude-
-    // side integration test in `tests/integration/mod.rs` only asserts the
-    // marker write, not the read-back through `read_session_history`.
-    setup_env_once();
-    let _guard = TEST_MUTEX.lock().unwrap();
-    reset_session_files();
-
-    let tmp = tempfile::tempdir().unwrap();
-    let history = tmp.path().join("session.jsonl");
-    std::fs::write(&history, b"line1\nline2\n").unwrap();
-    let path_file = tmp.path().join("path.txt");
-    std::fs::write(&path_file, history.to_string_lossy().as_bytes()).unwrap();
-
-    let bytes =
-        guest_agent::session_history::read_session_history(path_file.to_str().unwrap()).unwrap();
-    assert_eq!(bytes, b"line1\nline2\n");
 }
